Cleaning done. Added UART.h library with UART functions.
